Calm down a warning in 1d about "(unsigned < 0) is always false". This code cannot be reached in 1d.
git-svn-id: https://svn.dealii.org/trunk@6734 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/fe/fe_q.cc
@@ -1436,7 +1436,8 @@ FE_Q<dim>::has_support_on_face (const unsigned int shape_index,
         {
           const unsigned int quad_index 
             = (shape_index - first_quad_index) / dofs_per_quad;
-          Assert (quad_index < GeometryInfo<dim>::quads_per_cell,
+          Assert (static_cast<signed int>(quad_index) <
+                  static_cast<signed int>(GeometryInfo<dim>::quads_per_cell),
                   ExcInternalError());
           
                                            // in 2d, cell bubble are
